requirements.txt
absl-py==1.1.0
astunparse==1.6.3
autopep8==1.6.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask==2.1.2
flatbuffers==1.12
gast==0.4.0
google-auth==2.7.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
gunicorn==20.1.0
h5py==3.7.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
mysql-connector==2.2.9
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.1.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.9
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
